requirements.txt
annotated-types==0.7.0
black==25.1.0
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
isort==6.0.1
kiwisolver==1.4.8
matplotlib==3.10.3
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.1
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.13.2
